% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate.R
\name{.interpolate2}
\alias{.interpolate2}
\title{interpolate2}
\usage{
.interpolate2(zI, gap, blocks, delT, sigClip, freqSave, progress)
}
\arguments{
\item{zI}{}

\item{gap}{}

\item{blocks}{}

\item{delT}{}

\item{sigClip}{}

\item{freqSave}{}

\item{progress}{}
}
\description{
interpolate2
}
