% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMt.R
\name{estimateMt}
\alias{estimateMt}
\title{Title estimateMt}
\usage{
estimateMt(x, N, nw, k, pMax)
}
\arguments{
\item{x}{data?}

\item{N}{parameter for fitting quadratic}

\item{nw}{parameter for fitting quadratic}

\item{k}{parameter for fitting quadratic}

\item{pMax}{parameter for fitting quadratic}
}
\value{
phat
}
\description{
Title estimateMt
}
