% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTt.R
\name{estimateTt}
\alias{estimateTt}
\title{Estimate the sinusoidal components in a time series.}
\usage{
estimateTt(x, epsilon, dT, nw, k, sigClip, progress = FALSE, freqIn = NULL)
}
\arguments{
\item{x}{The input time series.}

\item{epsilon}{A tolerance for peak estimation convergence.}

\item{dT}{The time interval between data points.}

\item{nw}{The time-half bandwidth parameter for multitaper.}

\item{k}{The number of tapers.}

\item{sigClip}{Significance level for peak detection.}

\item{progress}{If TRUE, show progress information.}

\item{freqIn}{A pre-defined set of frequencies (optional).}
}
\value{
A matrix containing sinusoidal components.
}
\description{
This function estimates sinusoidal components in a time series using the multitaper method.
}
\details{
The function follows these algorithmic steps:
1. Spectrum/F-test pilot estimate using the multitaper method.
2. Estimation of significant peaks based on the significance level (sigClip).
3. Refinement of peak locations by eliminating duplicates and optimizing peak frequencies.
4. Estimation of phase and amplitude by inverting the spectrum (line component removal).
}
\examples{
estimateTt(x, epsilon = 1e-10, dT = 0.01, nw = 4, k = 5, sigClip = 0.05)

}
