% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateWt.R
\name{estimateWt}
\alias{estimateWt}
\title{Utility that takes pre-interpolated (linear?) series and computes spectra
and ACVF / CCVF series from them, which are then passed to mwXSwiener.}
\usage{
estimateWt(xd1, xd2, ok1, ok2, dT, blocks, neh, maxlag, clipMax)
}
\arguments{
\item{xd1}{}

\item{xd2}{}

\item{ok1}{}

\item{ok2}{}

\item{dT}{}

\item{blocks}{}

\item{neh}{}

\item{maxlag}{}

\item{clipMax}{}
}
\description{
Utility that takes pre-interpolated (linear?) series and computes spectra
and ACVF / CCVF series from them, which are then passed to mwXSwiener.
}
