% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolate.R
\name{interpolate}
\alias{interpolate}
\title{interpolate}
\usage{
interpolate(z, gap, maxit = 20, progress = FALSE, sigClip = 0.999, delT = 1)
}
\arguments{
\item{z}{time series with gaps, denoted by \code{NA}.}

\item{gap}{indexes of missing values, from \code{1:N}, where \code{N = length(z)}.}

\item{maxit}{maximum number of iterations for convergence in interpolation.}

\item{progress}{logical: should progress be written to screen as iterations proceed?}

\item{sigClip}{probabilistic significance for choice of line components, dividing series
into ``signal'' and ``noise'' (see algorithm for more). Suggested that this be kept
above \code{0.95} at a minimum.}

\item{delT}{the time step delta-t in seconds.}
}
\value{
returns a list of the interpolated timeseries
}
\description{
Univariate interpolation of gappy time series.
}
\examples{
library("tsinterp")
         data("flux")
         z1 <- flux$SagOrig
         z1[which(flux$S == FALSE)] <- NA
         interpolate(z, gap, maxit = 20, progress=FALSE, sigClip=0.999, delT=1)




}
