% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePeriod.R
\name{removePeriod}
\alias{removePeriod}
\title{Removes period from a data for time series interpolation purpose}
\usage{
removePeriod(xd, f0, nw, k, deltaT, warn = FALSE, prec = 1e-10, sigClip)
}
\arguments{
\item{xd}{: data}

\item{f0}{: freq of periodicity to remove}

\item{nw}{: parameters of multitaper}

\item{k}{: parameters of multitaper}

\item{deltaT}{: parameter of xd}

\item{warn}{}

\item{prec}{: starting precision for finding a good nFFT for removal}

\item{sigClip}{}
}
\value{
inv
}
\description{
Removes period from a data for time series interpolation purpose
}
