\name{tsinterp-package}
\alias{tsinterp-package}
\alias{tsinterp}
\docType{package}
\title{
    tsinterp: A Time Series Interpolation Package
}
\description{
    Contains implementations of time series interpolation algorithms developed
    by Wesley Burr as part of his PhD.
}
\details{
\tabular{ll}{
Package: \tab tsinterp\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2014-09-24\cr
License: \tab GPL-2 \cr
}
}
\author{
Wesley Burr, with contributions from LINPACK

Maintainer: Wesley Burr <wesley.burr@gmail.com>
}
\references{
  Thomson, D.J (1982)
  Spectrum estimation and harmonic analysis. \emph{Proceedings of the IEEE}
  Volume \bold{70}, number 9, pp. 1055--1096.

  Burr, W.S. (2012)
  Air Pollution and Health: Time Series Tools and Analysis. PhD Thesis,
  Queen's University at Kingston, October, 2012. https://qspace.library.queensu.ca/handle/1974/7617

}
\keyword{time series}
\keyword{interpolation}
\keyword{ts}
\keyword{Slepian}
\seealso{
\code{\link[multitaper:spec.mtm]{multitaper}}
}

