% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tsinterp}
\alias{tsinterp}
\title{tsinterp: A Time Series Interpolation Package}
\description{
Contains implementations of time series interpolation algorithms developed
by Wesley Burr as part of his PhD.
}
\references{
Thomson, D.J (1982)
Spectrum estimation and harmonic analysis. \emph{Proceedings of the IEEE}
Volume \bold{70}, number 9, pp. 1055--1096.

Burr, W.S. (2012)
Air Pollution and Health: Time Series Tools and Analysis. PhD Thesis,
Queen's University at Kingston, October, 2012. https://qspace.library.queensu.ca/handle/1974/7617
}
\author{
[aut, cre] Wesley S. Burr `<wesley.burr@gmail.com>`, with contributions from LINPACK
}
\keyword{Slepian}
\keyword{interpolation}
\keyword{series}
\keyword{time}
\keyword{ts}
