% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBlocks.R
\name{findBlocks}
\alias{findBlocks}
\title{findBlocks
Find blocks of missing points from a mask of a time series
Scans element-wise to find blocks of missing points}
\usage{
findBlocks(mask)
}
\arguments{
\item{mask}{Usage: findBlocks(mask)}
}
\value{
a matrix of missing points as blocks, size \code{M * 3}. The start-point, end-point and length of each block are recorded row-wise
}
\description{
findBlocks
Find blocks of missing points from a mask of a time series
Scans element-wise to find blocks of missing points
}
\examples{
library("tsinterp") 
          data("flux")
          miss <- flux$S 
          miss[miss == FALSE] <- NA
          blocks <- findBlocks(miss)

}
