% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_m_t.R
\name{gen_m_t}
\alias{gen_m_t}
\title{gen_m_t is a used to generate mt for the interpolation function}
\usage{
gen_m_t(zI, N, delT, sigClip, progress)
}
\arguments{
\item{zI}{Input time series; should have no missing values.}

\item{N}{Length of zI}

\item{delT}{delta-T; the time step of the time series.}

\item{sigClip}{Significance level for periodic component detection.}

\item{progress}{Should a progress tick be printed or not?}
}
\value{
a class list that has Ttp and freqsave as objects
}
\description{
gen_m_t is a used to generate mt for the interpolation function
}
\details{
A function that, given argument zI, computes a MtP/TtP
pairing of estimates, which can then be used in the 
maximum-likelihood iterative step of the EM algorithm
for interpolation.
}
